feat: Implement taggleLinks function to manage link interaction state
