+++ styles/content.css
@@ -7,9 +7,11 @@
   pointer-events: all !important;
 }
 
-.disable-links a {
+.disable-links a,
+.disable-links button,
+.disable-links [onclick],
+.disable-links [role="link"] {
   pointer-events: none !important;
-  cursor: crosshair !important; /* تغییر نشانگر موس */
-  color: inherit; /* اخیاری: برای حذف رنگ پیش‌فرض لینک‌ها */
-  text-decoration: none;
+  /* cursor: crosshair !important; */
+  color: inherit;
 }
